refactor(storage): resolve the storage driver via @ext/storage-driver
Static seam like the other ext backends: the local-disk driver moves to
src/ext-default/storage-driver.ts and getStorage() falls back to it,
with registerStorageDriver kept as a runtime override hook. Removes the
need for shells to register a driver at runtime (Workers builds have no
reliable instrumentation hook, and a registration side-effect module
created an import cycle through the scope resolver).
